requirements.txt
playwright
beautifulsoup4
requests
pymongo
python-dotenv
chromadb
langgraph